- Project reorganization - Directory.Build.props - Reintroducing namespaces and class names to program.cs in the workers - Imports => WorkerContext
