+++ Blazor.WebWorkers.sln
@@ -1,15 +1,26 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
+# 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KristofferStrube.Blazor.WebWorkers.WasmExample", "samples\KristofferStrube.Blazor.WebWorkers.WasmExample\KristofferStrube.Blazor.WebWorkers.WasmExample.csproj", "{47569989-372C-4BAD-8922-9576F40D7613}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KristofferStrube.Blazor.WebWorkers.Samples.WasmExample", "samples\KristofferStrube.Blazor.WebWorkers.Samples.WasmExample\KristofferStrube.Blazor.WebWorkers.Samples.WasmExample.csproj", "{47569989-372C-4BAD-8922-9576F40D7613}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KristofferStrube.Blazor.WebWorkers.StringSumWorker", "samples\KristofferStrube.Blazor.WebWorkers.StringSumWorker\KristofferStrube.Blazor.WebWorkers.StringSumWorker.csproj", "{8A5BB220-5A13-4562-8895-C11F2D340FFC}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KristofferStrube.Blazor.WebWorkers.Samples.StringSumWorker", "samples\KristofferStrube.Blazor.WebWorkers.Samples.StringSumWorker\KristofferStrube.Blazor.WebWorkers.Samples.StringSumWorker.csproj", "{8A5BB220-5A13-4562-8895-C11F2D340FFC}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KristofferStrube.Blazor.WebWorkers", "src\KristofferStrube.Blazor.WebWorkers\KristofferStrube.Blazor.WebWorkers.csproj", "{0E04FCA1-8997-45DC-A195-ABBCC6CB0265}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KristofferStrube.Blazor.WebWorkers.PongWorker", "samples\KristofferStrube.Blazor.WebWorkers.PongWorker\KristofferStrube.Blazor.WebWorkers.PongWorker.csproj", "{2EDD6A12-7511-4278-B5B3-6898BB1E4BF4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KristofferStrube.Blazor.WebWorkers.Samples.PongWorker", "samples\KristofferStrube.Blazor.WebWorkers.Samples.PongWorker\KristofferStrube.Blazor.WebWorkers.Samples.PongWorker.csproj", "{2EDD6A12-7511-4278-B5B3-6898BB1E4BF4}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{DB025BBF-9EA0-42EE-9C74-B58964D32362}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Real World #1", "Real World #1", "{490E7C4D-597C-47EF-9D73-894CAF2A906D}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{F0743A97-325E-4A45-832B-55948952D6A0}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{F36452E6-F33A-4301-BE92-18A42B1F6E2E}"
+	ProjectSection(SolutionItems) = preProject
+		Directory.Build.props = Directory.Build.props
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -36,6 +47,13 @@ Global
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{47569989-372C-4BAD-8922-9576F40D7613} = {DB025BBF-9EA0-42EE-9C74-B58964D32362}
+		{8A5BB220-5A13-4562-8895-C11F2D340FFC} = {DB025BBF-9EA0-42EE-9C74-B58964D32362}
+		{0E04FCA1-8997-45DC-A195-ABBCC6CB0265} = {F0743A97-325E-4A45-832B-55948952D6A0}
+		{2EDD6A12-7511-4278-B5B3-6898BB1E4BF4} = {DB025BBF-9EA0-42EE-9C74-B58964D32362}
+		{490E7C4D-597C-47EF-9D73-894CAF2A906D} = {DB025BBF-9EA0-42EE-9C74-B58964D32362}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {B314BCF6-081C-4A66-985B-BD5DF3CB4633}
