--- samples/KristofferStrube.Blazor.WebWorkers.PongWorker/AssemblyPongWorker.cs
@@ -1,5 +1,0 @@
-﻿namespace KristofferStrube.Blazor.WebWorkers.PongWorker;
-
-public class AssemblyPongWorker
-{
-}
