+++ samples/KristofferStrube.Blazor.WebWorkers.Samples.PongWorker/AssemblyPongWorker.cs
@@ -1,0 +1,5 @@
+﻿namespace KristofferStrube.Blazor.WebWorkers.Samples.PongWorker;
+
+public class AssemblyPongWorker
+{
+}
