+++ samples/KristofferStrube.Blazor.WebWorkers.Samples.PongWorker/Properties/launchSettings.json
