+++ samples/KristofferStrube.Blazor.WebWorkers.Samples.StringSumWorker/Properties/launchSettings.json
