+++ samples/KristofferStrube.Blazor.WebWorkers.Samples.WasmExample/App.razor
