+++ samples/KristofferStrube.Blazor.WebWorkers.Samples.WasmExample/KristofferStrube.Blazor.WebWorkers.Samples.WasmExample.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.BlazorWebAssembly">
+
+	<PropertyGroup>
+		<TargetFramework>net8.0</TargetFramework>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="8.*" />
+		<PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.DevServer" Version="8.*" PrivateAssets="all" />
+		<PackageReference Include="System.Text.Json" Version="8.*" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<ProjectReference Include="..\..\src\KristofferStrube.Blazor.WebWorkers\KristofferStrube.Blazor.WebWorkers.csproj" />
+		<ProjectReference Include="..\KristofferStrube.Blazor.WebWorkers.Samples.PongWorker\KristofferStrube.Blazor.WebWorkers.Samples.PongWorker.csproj" />
+		<ProjectReference Include="..\KristofferStrube.Blazor.WebWorkers.Samples.StringSumWorker\KristofferStrube.Blazor.WebWorkers.Samples.StringSumWorker.csproj" />
+	</ItemGroup>
+
+</Project>
