+++ samples/KristofferStrube.Blazor.WebWorkers.Samples.WasmExample/Layout/NavMenu.razor.css
