+++ samples/KristofferStrube.Blazor.WebWorkers.Samples.WasmExample/Pages/Home.razor
@@ -1,7 +1,7 @@
 ﻿@page "/"
 @using KristofferStrube.Blazor.DOM
 @using KristofferStrube.Blazor.Window
-@using KristofferStrube.Blazor.WebWorkers.StringSumWorker
+@using KristofferStrube.Blazor.WebWorkers.Samples.StringSumWorker
 @using System.Text.Json
 @using System.Diagnostics
 @implements IAsyncDisposable
