+++ samples/KristofferStrube.Blazor.WebWorkers.Samples.WasmExample/_Imports.razor
@@ -6,5 +6,5 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
-@using KristofferStrube.Blazor.WebWorkers.WasmExample
-@using KristofferStrube.Blazor.WebWorkers.WasmExample.Layout
+@using KristofferStrube.Blazor.WebWorkers.Samples.WasmExample
+@using KristofferStrube.Blazor.WebWorkers.Samples.WasmExample.Layout
