+++ samples/KristofferStrube.Blazor.WebWorkers.Samples.WasmExample/wwwroot/404.html
