+++ samples/KristofferStrube.Blazor.WebWorkers.Samples.WasmExample/wwwroot/favicon.png
