+++ samples/KristofferStrube.Blazor.WebWorkers.Samples.WasmExample/wwwroot/icon-192.png
