+++ samples/KristofferStrube.Blazor.WebWorkers.Samples.WasmExample/wwwroot/index.html
