+++ samples/KristofferStrube.Blazor.WebWorkers.Samples.WasmExample/wwwroot/stringSumWorker.js
