--- samples/KristofferStrube.Blazor.WebWorkers.StringSumWorker/Program.cs
@@ -1,6 +1,0 @@
-﻿using KristofferStrube.Blazor.WebWorkers.StringSumWorker;
-
-if (!OperatingSystem.IsBrowser())
-    throw new PlatformNotSupportedException("Can only be run in the browser!");
-
-await new StringSumJob().StartAsync();
