--- samples/KristofferStrube.Blazor.WebWorkers.WasmExample/KristofferStrube.Blazor.WebWorkers.WasmExample.csproj
@@ -1,20 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.BlazorWebAssembly">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="8.0.3" />
-    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.DevServer" Version="8.0.3" PrivateAssets="all" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\KristofferStrube.Blazor.WebWorkers\KristofferStrube.Blazor.WebWorkers.csproj" />
-    <ProjectReference Include="..\KristofferStrube.Blazor.WebWorkers.PongWorker\KristofferStrube.Blazor.WebWorkers.PongWorker.csproj" />
-    <ProjectReference Include="..\KristofferStrube.Blazor.WebWorkers.StringSumWorker\KristofferStrube.Blazor.WebWorkers.StringSumWorker.csproj" />
-  </ItemGroup>
-
-</Project>
