+++ src/KristofferStrube.Blazor.WebWorkers/Jobs/IJob.cs
