+++ src/KristofferStrube.Blazor.WebWorkers/Options/WorkerType.cs
@@ -1,20 +1,22 @@
-﻿using KristofferStrube.Blazor.WebWorkers.Converters;
-using System.Text.Json.Serialization;
+﻿using System.Text.Json.Serialization;
 
 namespace KristofferStrube.Blazor.WebWorkers;
 
 /// <summary>
 /// The type of worker.
 /// </summary>
-[JsonConverter(typeof(WorkerTypeConverter))]
+[JsonConverter(typeof(JsonStringEnumConverter<WorkerType>))]
 public enum WorkerType
 {
+
     /// <summary>
     /// A classic worker.
     /// </summary>
     Classic,
+
     /// <summary>
     /// A worker that is also a module.
     /// </summary>
     Module
+
 }
