+++ src/KristofferStrube.Blazor.WebWorkers/Workers/SlimWorker.cs
