+++ src/KristofferStrube.Blazor.WebWorkers/Workers/Worker.cs
